Test methods moved to ActionEvaluatorTest
Test method moved to ActionEvaluatorTest
