+++ Libplanet.Tests/Blockchain/BlockChainTest.cs
@@ -1463,76 +1463,6 @@ namespace Libplanet.Tests.Blockchain
         }
 
         [Fact]
-        public void EvaluatePolicyBlockAction()
-        {
-            (var addresses, Transaction<DumbAction>[] txs) =
-                MakeFixturesForAppendTests();
-            var genesis = _blockChain.Genesis;
-
-            Block<DumbAction> block1 = TestUtils.MineNext(
-                genesis,
-                txs,
-                difficulty: _blockChain.Policy.GetNextBlockDifficulty(_blockChain)
-            ).AttachStateRootHash(_blockChain.StateStore, _policy.BlockAction);
-
-            var miner = genesis.Miner.GetValueOrDefault();
-            var stateCompleterSet = StateCompleterSet<DumbAction>.Recalculate;
-            AccountStateGetter accountStateGetter = address => _blockChain.GetState(
-                address,
-                block1.PreviousHash,
-                stateCompleterSet.StateCompleter);
-            AccountBalanceGetter accountBalanceGetter =
-                (address, currency) => _blockChain.GetBalance(
-                    address,
-                    currency,
-                    block1.PreviousHash,
-                    stateCompleterSet.FungibleAssetStateCompleter);
-            IAccountStateDelta previousStates = block1.ProtocolVersion > 0
-                    ? new AccountStateDeltaImpl(
-                        accountStateGetter, accountBalanceGetter, block1.Miner.GetValueOrDefault())
-                    : new AccountStateDeltaImplV0(
-                        accountStateGetter, accountBalanceGetter, block1.Miner.GetValueOrDefault());
-
-            var policyBlockActionEvaluation = _blockChain.ActionEvaluator.EvaluatePolicyBlockAction(
-                block1,
-                previousStates,
-                null);
-            Assert.Equal(_blockChain.Policy.BlockAction, policyBlockActionEvaluation.Action);
-            Assert.Equal(
-                (Integer)2,
-                (Integer)policyBlockActionEvaluation.OutputStates.GetState(miner));
-            Assert.True(policyBlockActionEvaluation.InputContext.BlockAction);
-
-            _blockChain.ExecuteActions(block1);
-            _blockChain.Append(
-                block1,
-                DateTimeOffset.UtcNow,
-                evaluateActions: false,
-                renderBlocks: true,
-                renderActions: false);
-            previousStates = block1.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    address => _blockChain.GetState(address, block1.PreviousHash),
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    block1.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    address => _blockChain.GetState(address, block1.PreviousHash),
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    block1.Miner.GetValueOrDefault());
-            var txEvaluations = _blockChain.ActionEvaluator.EvaluateTxsGradually(
-                block1,
-                previousStates).Select(te => te.Item2).ToList();
-            previousStates = txEvaluations.Last().OutputStates;
-            policyBlockActionEvaluation = _blockChain.ActionEvaluator.EvaluatePolicyBlockAction(
-                block1,
-                previousStates,
-                null);
-            Assert.Equal(
-                (Integer)2,
-                (Integer)policyBlockActionEvaluation.OutputStates.GetState(miner));
-        }
-
-        [Fact]
         public async void BlockActionWithMultipleAddress()
         {
             var miner0 = _blockChain.Genesis.Miner;
