+++ Libplanet.Tests/Blocks/BlockTest.cs
@@ -1,17 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Linq;
-using System.Numerics;
 using Bencodex;
 using Bencodex.Types;
 using Libplanet.Action;
-using Libplanet.Assets;
 using Libplanet.Blocks;
 using Libplanet.Crypto;
-using Libplanet.Tests.Action;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tx;
 using Xunit;
@@ -199,287 +195,6 @@ namespace Libplanet.Tests.Blocks
             Assert.Equal(actual.BytesLength, encoded.Length);
             AssertBytesEqual(_fx.HasTx.Serialize(), encoded);
             Assert.Equal(_fx.HasTx.BytesLength, encoded.Length);
-        }
-
-        [SuppressMessage(
-            "Microsoft.StyleCop.CSharp.ReadabilityRules",
-            "SA1118",
-            Justification = "Long array literals should be multiline.")]
-        [Fact]
-        public void EvaluateActionsPerTx()
-        {
-            Address[] addresses =
-            {
-                _fx.TxFixture.Address1,
-                _fx.TxFixture.Address2,
-                _fx.TxFixture.Address3,
-                _fx.TxFixture.Address4,
-                _fx.TxFixture.Address5,
-            };
-            DumbAction MakeAction(Address address, char identifier, Address? transferTo = null) =>
-                new DumbAction(
-                    targetAddress: address,
-                    item: identifier.ToString(),
-                    recordRehearsal: false,
-                    recordRandom: true,
-                    transfer: transferTo is Address to
-                        ? Tuple.Create<Address, Address, BigInteger>(address, to, 5)
-                        : null);
-            Block<DumbAction> genesis = MineGenesis<DumbAction>();
-            ActionEvaluator<DumbAction> actionEvaluator = new ActionEvaluator<DumbAction>(
-                policyBlockAction: null,
-                stateGetter: ActionEvaluator<DumbAction>.NullStateGetter,
-                balanceGetter: ActionEvaluator<DumbAction>.NullBalanceGetter,
-                trieGetter: null);
-            IAccountStateDelta previousStates = genesis.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    genesis.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    genesis.Miner.GetValueOrDefault());
-            Assert.Empty(
-                actionEvaluator.EvaluateTxsGradually(
-                    genesis,
-                    previousStates));
-
-            Transaction<DumbAction>[] blockIdx1Txs =
-            {
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey1,
-                    _fx.Genesis.Hash,
-                    new[]
-                    {
-                        MakeAction(addresses[0], 'A', addresses[1]),
-                        MakeAction(addresses[1], 'B', addresses[2]),
-                    },
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(1)),
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey2,
-                    _fx.Genesis.Hash,
-                    new[] { MakeAction(addresses[2], 'C', addresses[3]) },
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(2)),
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey3,
-                    _fx.Genesis.Hash,
-                    new DumbAction[0],
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(8)),
-            };
-            int i = 0;
-            foreach (Transaction<DumbAction> tx in blockIdx1Txs)
-            {
-                _output.WriteLine("{0}[{1}] = {2}", nameof(blockIdx1Txs), i, tx.Id);
-            }
-
-            Block<DumbAction> blockIdx1 = MineNext(genesis, blockIdx1Txs, new byte[] { });
-            previousStates = blockIdx1.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    blockIdx1.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    blockIdx1.Miner.GetValueOrDefault());
-            var pairs = actionEvaluator.EvaluateTxsGradually(
-                blockIdx1,
-                previousStates).ToImmutableArray();
-            int randomValue = 0;
-            (int, int, string[], Address)[] expectations =
-            {
-                (1, 0, new[] { null, null, "C", null, null }, _fx.TxFixture.Address2),
-                (0, 0, new[] { "A", null, "C", null, null }, _fx.TxFixture.Address1),
-                (0, 1, new[] { "A", "B", "C", null, null }, _fx.TxFixture.Address1),
-            };
-            Assert.Equal(expectations.Length, pairs.Length);
-            foreach (var (expect, pair) in expectations.Zip(pairs, ValueTuple.Create))
-            {
-                ActionEvaluation eval = pair.Item2;
-                Assert.Equal(blockIdx1Txs[expect.Item1], pair.Item1);
-                Assert.Equal(blockIdx1Txs[expect.Item1].Actions[expect.Item2], eval.Action);
-                Assert.Equal(expect.Item4, eval.InputContext.Signer);
-                Assert.Equal(GenesisMinerAddress, eval.InputContext.Miner);
-                Assert.Equal(blockIdx1.Index, eval.InputContext.BlockIndex);
-                Assert.False(eval.InputContext.Rehearsal);
-                randomValue = eval.InputContext.Random.Next();
-                Assert.Equal(
-                    (Integer)eval.OutputStates.GetState(
-                        DumbAction.RandomRecordsAddress),
-                    (Integer)randomValue);
-                Assert.Equal(
-                    expect.Item3,
-                    addresses.Select(eval.OutputStates.GetState)
-                        .Select(x => x is Text t ? t.Value : null));
-            }
-
-            previousStates = blockIdx1.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    blockIdx1.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    ActionEvaluator<DumbAction>.NullAccountStateGetter,
-                    ActionEvaluator<DumbAction>.NullAccountBalanceGetter,
-                    blockIdx1.Miner.GetValueOrDefault());
-            ActionEvaluation[] evals1 = actionEvaluator.EvaluateBlock(
-                blockIdx1,
-                DateTimeOffset.UtcNow,
-                previousStates).ToArray();
-            IImmutableDictionary<Address, IValue> dirty1 = evals1.GetDirtyStates();
-            IImmutableDictionary<(Address, Currency), FungibleAssetValue> balances1 =
-                evals1.GetDirtyBalances();
-            Assert.Equal(
-                new Dictionary<Address, IValue>
-                {
-                    [addresses[0]] = (Text)"A",
-                    [addresses[1]] = (Text)"B",
-                    [addresses[2]] = (Text)"C",
-                    [DumbAction.RandomRecordsAddress] = (Integer)randomValue,
-                }.ToImmutableDictionary(),
-                dirty1);
-            Assert.Equal(
-                new Dictionary<(Address, Currency), FungibleAssetValue>
-                {
-                    [(addresses[0], DumbAction.DumbCurrency)] =
-                        new FungibleAssetValue(DumbAction.DumbCurrency, -5, 0),
-                    [(addresses[1], DumbAction.DumbCurrency)] =
-                        new FungibleAssetValue(DumbAction.DumbCurrency),
-                    [(addresses[2], DumbAction.DumbCurrency)] =
-                        new FungibleAssetValue(DumbAction.DumbCurrency),
-                    [(addresses[3], DumbAction.DumbCurrency)] =
-                        new FungibleAssetValue(DumbAction.DumbCurrency, 5, 0),
-                }.ToImmutableDictionary(),
-                balances1);
-
-            Transaction<DumbAction>[] blockIdx2Txs =
-            {
-                // Note that these timestamps in themselves does not have any meanings but are
-                // only arbitrary.  These purpose to make their evaluation order in a block
-                // equal to the order we (the test) intend:
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey1,
-                    _fx.Genesis.Hash,
-                    new[] { MakeAction(addresses[0], 'D') },
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(1)),
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey2,
-                    _fx.Genesis.Hash,
-                    new[] { MakeAction(addresses[3], 'E') },
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(3)),
-                Transaction<DumbAction>.Create(
-                    0,
-                    _fx.TxFixture.PrivateKey3,
-                    _fx.Genesis.Hash,
-                    new[]
-                    {
-                        new DumbAction(
-                            addresses[4],
-                            "RecordRehearsal",
-                            transferFrom: addresses[0],
-                            transferTo: addresses[4],
-                            transferAmount: 8,
-                            recordRehearsal: true,
-                            recordRandom: true),
-                    },
-                    timestamp: DateTimeOffset.MinValue.AddSeconds(5)),
-            };
-            i = 0;
-            foreach (Transaction<DumbAction> tx in blockIdx2Txs)
-            {
-                _output.WriteLine("{0}[{1}] = {2}", nameof(blockIdx2Txs), i, tx.Id);
-            }
-
-            Block<DumbAction> blockIdx2 = MineNext(blockIdx1, blockIdx2Txs, new byte[] { });
-            AccountStateGetter accountStateGetter = dirty1.GetValueOrDefault;
-            AccountBalanceGetter accountBalanceGetter = (address, currency)
-                => balances1.TryGetValue((address, currency), out FungibleAssetValue v)
-                    ? v
-                    : new FungibleAssetValue(currency);
-            previousStates = blockIdx2.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    accountStateGetter,
-                    accountBalanceGetter,
-                    blockIdx2.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    accountStateGetter,
-                    accountBalanceGetter,
-                    blockIdx2.Miner.GetValueOrDefault());
-            pairs = actionEvaluator.EvaluateTxsGradually(
-                blockIdx2,
-                previousStates).ToImmutableArray();
-            expectations = new[]
-            {
-                (0, 0, new[] { "A,D", "B", "C", null, null }, _fx.TxFixture.Address1),
-                (1, 0, new[] { "A,D", "B", "C", "E", null }, _fx.TxFixture.Address2),
-                (
-                    2,
-                    0,
-                    new[] { "A,D", "B", "C", "E", "RecordRehearsal:False" },
-                    _fx.TxFixture.Address3
-                ),
-            };
-            Assert.Equal(expectations.Length, pairs.Length);
-            foreach (var (expect, pair) in expectations.Zip(pairs, ValueTuple.Create))
-            {
-                ActionEvaluation eval = pair.Item2;
-                Assert.Equal(blockIdx2Txs[expect.Item1], pair.Item1);
-                Assert.Equal(
-                    blockIdx2Txs[expect.Item1].Actions[expect.Item2],
-                    eval.Action);
-                Assert.Equal(expect.Item4, eval.InputContext.Signer);
-                Assert.Equal(GenesisMinerAddress, eval.InputContext.Miner);
-                Assert.Equal(blockIdx2.Index, eval.InputContext.BlockIndex);
-                Assert.False(eval.InputContext.Rehearsal);
-                Assert.Null(eval.Exception);
-                Assert.Equal(
-                    expect.Item3,
-                    addresses
-                        .Select(eval.OutputStates.GetState)
-                        .Select(x => x is Text t ? t.Value : null));
-                randomValue = eval.InputContext.Random.Next();
-                Assert.Equal(
-                    eval.OutputStates.GetState(
-                        DumbAction.RandomRecordsAddress),
-                    (Integer)randomValue);
-            }
-
-            accountStateGetter = dirty1.GetValueOrDefault;
-            accountBalanceGetter = (address, currency) => balances1.TryGetValue(
-                (address, currency), out FungibleAssetValue value)
-                    ? value
-                    : new FungibleAssetValue(currency);
-            previousStates = blockIdx2.ProtocolVersion > 0
-                ? new AccountStateDeltaImpl(
-                    accountStateGetter,
-                    accountBalanceGetter,
-                    blockIdx2.Miner.GetValueOrDefault())
-                : new AccountStateDeltaImplV0(
-                    accountStateGetter,
-                    accountBalanceGetter,
-                    blockIdx2.Miner.GetValueOrDefault());
-            var evals2 = actionEvaluator.EvaluateBlock(
-                blockIdx2,
-                DateTimeOffset.UtcNow,
-                previousStates).ToArray();
-            IImmutableDictionary<Address, IValue> dirty2 = evals2.GetDirtyStates();
-            IImmutableDictionary<(Address, Currency), FungibleAssetValue> balances2 =
-                evals2.GetDirtyBalances();
-            Assert.Equal(
-                new Dictionary<Address, IValue>
-                {
-                    [addresses[0]] = (Text)"A,D",
-                    [addresses[3]] = (Text)"E",
-                    [addresses[4]] = (Text)"RecordRehearsal:False",
-                    [DumbAction.RandomRecordsAddress] = (Integer)randomValue,
-                }.ToImmutableDictionary(),
-                dirty2);
         }
 
         [Fact]
